Integrate (SimpleSaml) Authorisation
This PR provides the admin panel with authorization. Authorization can be given in 2 ways:
1) via a federated login (stepup-saml) and
2) via a local login.
The federated login has yet to be tested.
The setup is now that an admin (contact) can manage 1 or more realm configurations. A super admin (marked as super in contact database) can manage all realms.
The migrations of the contact entities cannot yet be performed (and are therefore not included) due to the enum options in other entities. This has yet to be resolved.